// examples/petstore/generated/api/models/pet.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Petstore API
 * A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
 * OpenAPI spec version: 1.0.0
 */
import type { Category } from './category';
import type { PetStatus } from './petStatus';
import type { Tag } from './tag';

export interface Pet {
  category?: Category;
  /** When the pet was created */
  createdAt?: string;
  /** Unique identifier for the pet */
  id: number;
  /** Name of the pet */
  name: string;
  /** List of photo URLs */
  photoUrls?: string[];
  /** Pet status in the store */
  status?: PetStatus;
  /** List of tags */
  tags?: Tag[];
  /** When the pet was last updated */
  updatedAt?: string;
}