// examples/petstore/generated/api/models/updatePet.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Petstore API
 * A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
 * OpenAPI spec version: 1.0.0
 */
import type { Category } from './category';
import type { UpdatePetStatus } from './updatePetStatus';
import type { Tag } from './tag';

export interface UpdatePet {
  category?: Category;
  /** Name of the pet */
  name?: string;
  /** List of photo URLs */
  photoUrls?: string[];
  /** Pet status in the store */
  status?: UpdatePetStatus;
  /** List of tags */
  tags?: Tag[];
}