// examples/petstore/generated/api/models/user.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Petstore API
 * A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
 * OpenAPI spec version: 1.0.0
 */
import type { UserUserStatus } from './userUserStatus';

export interface User {
  /** When the user was created */
  createdAt?: string;
  /** User email */
  email: string;
  /** First name */
  firstName?: string;
  /** User ID */
  id: number;
  /** Last name */
  lastName?: string;
  /** User phone number */
  phone?: string;
  /** When the user was last updated */
  updatedAt?: string;
  /** User name */
  username: string;
  /** User Status */
  userStatus?: UserUserStatus;
}