// examples/petstore/generated/api/models/newUser.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Petstore API
 * A sample API that uses a petstore as an example to demonstrate features in the OpenAPI 3.0 specification
 * OpenAPI spec version: 1.0.0
 */
import type { NewUserUserStatus } from './newUserUserStatus';

export interface NewUser {
  /** User email */
  email: string;
  /** First name */
  firstName?: string;
  /** Last name */
  lastName?: string;
  /** User phone number */
  phone?: string;
  /** User name */
  username: string;
  /** User Status */
  userStatus?: NewUserUserStatus;
}